ChainRulesTestUtils

ChainRulesTestUtils.jl helps you test ChainRulesCore.frule and ChainRulesCore.rrule methods, when adding rules for your functions in your own packages. For information about ChainRules, including how to write rules, refer to the general ChainRules Documentation:

Canonical example

Let's suppose a custom transformation has been defined

function two2three(x1::Float64, x2::Float64)
    return 1.0, 2.0*x1, 3.0*x2
end

along with the frule

using ChainRulesCore

function ChainRulesCore.frule((Δf, Δx1, Δx2), ::typeof(two2three), x1, x2)
    y = two2three(x1, x2)
    ∂y = Tangent{Tuple{Float64, Float64, Float64}}(ZeroTangent(), 2.0*Δx1, 3.0*Δx2)
    return y, ∂y
end

and rrule

function ChainRulesCore.rrule(::typeof(two2three), x1, x2)
    y = two2three(x1, x2)
    function two2three_pullback(Ȳ)
        return (NO_FIELDS, 2.0*Ȳ[2], 3.0*Ȳ[3])
    end
    return y, two2three_pullback
end

The test_frule/test_rrule helper function compares the frule/rrule outputs to the gradients obtained by finite differencing. They can be used for any type and number of inputs and outputs.

Testing the frule

test_frule takes in the function f and the primal input x. The call will test the frule for function f at the point x in the domain. Keep this in mind when testing discontinuous rules for functions like ReLU, which should ideally be tested at both x being above and below zero.

julia> using ChainRulesTestUtils;

julia> test_frule(two2three, 3.33, -7.77);
Test Summary:                            | Pass  Total
test_frule: two2three on Float64,Float64 |    5      5

Testing the rrule

test_rrule takes in the function f, and primal inputsr x. The call will test the rrule for function f at the point x, and similarly to frule some rules should be tested at multiple points in the domain.

julia> test_rrule(two2three, 3.33, -7.77);
Test Summary:                            | Pass  Total
test_rrule: two2three on Float64,Float64 |    6      6

Scalar example

For functions with a single argument and a single output, such as e.g. ReLU,

function relu(x::Real)
    return max(0, x)
end

with the frule and rrule defined with the help of @scalar_rule macro

@scalar_rule relu(x::Real) x <= 0 ? zero(x) : one(x)

test_scalar function is provided to test both the frule and the rrule with a single call.

julia> test_scalar(relu, 0.5);
Test Summary:            | Pass  Total
test_scalar: relu at 0.5 |    7      7

julia> test_scalar(relu, -0.5);
Test Summary:             | Pass  Total
test_scalar: relu at -0.5 |    7      7

Specifying Tangents

test_frule and test_rrule allow you to specify the tangents used for testing. This is done by passing in x ⊢ Δx, where x is the primal and Δx is the tangent, in the place of the primal inputs. If this is not done the tangent will be automatically generated via FiniteDifferences.rand_tangent. A special case of this is that if you specify it as x ⊢ NoTangent() then finite differencing will not be used on that input. Similarly, by setting the output_tangent keyword argument, you can specify the tangent for the primal output.

This can be useful when the default provided FiniteDifferences.rand_tangent doesn't produce the desired tangent for your type. For example the default tangent for an Int is NoTangent(). Which is correct e.g. when the Int represents a discrete integer like in indexing. But if you are testing something where the Int is actually a special case of a real number, then you would want to specify the tangent as a Float64.

Care must be taken when manually specifying tangents. In particular, when specifying the input tangents to test_frule and the output tangent to test_rrule. As these tangents are used to seed the derivative computation. Inserting inappropriate zeros can thus hide errors.

Custom finite differencing

If a package is using a custom finite differencing method of testing the frules and rrules, test_approx function provides a convenient way of comparing various types of differentials.

It is effectively (a, b) -> @test isapprox(a, b), but it preprocesses thunks and ChainRules differential types ZeroTangent(), NoTangent(), and Tangent, such that the error messages are helpful.

For example,

test_approx((@thunk 2*2.0), 4.1)

shows both the expression and the evaluated thunks

   Expression: isapprox(actual, expected; kwargs...)
   Evaluated: isapprox(4.0, 4.1)
ERROR: There was an error during testing

compared to

julia> @test isapprox(@thunk 2*2.0, 4.0)
Test Failed at REPL[52]:1
  Expression: isapprox(#= REPL[52]:1 =# @thunk((2 * 2.0, 4.0)))
   Evaluated: isapprox(Thunk(var"#24#25"()))
ERROR: There was an error during testing

which should have passed the test.

API Documentation

ChainRulesTestUtils.:⊢ — Type

primal ⊢ tangent

Infix shorthand method to construct a PrimalAndTangent. Enter via \vdash + tab on supporting editors.

ChainRulesTestUtils.TestIterator — Type

TestIterator{T,IS<:Base.IteratorSize,IE<:Base.IteratorEltype}

A configurable iterator for testing purposes.

TestIterator(data, itersize, itereltype)
TestIterator(data)

The iterator wraps another iterator data, such as an array, that must have at least as many features implemented as the test iterator and have a FiniteDifferences.to_vec overload. By default, the iterator it has the same features as data.

The optional methods eltype, length, and size are automatically defined and forwarded to data if the type arguments indicate that they should be defined.

ChainRulesTestUtils.test_approx — Function

test_approx(actual, expected, [msg]; kwargs...)

@test's that actual ≈ expected, but breaks up data such that human readable results are shown on failures. Understands things like unthunking ChainRuleCore.Thunks, etc.

If provided msg is printed on a failure. Often additional items are appended to msg to give bread-crumbs into nested structures.

All keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_frule — Method

test_frule(f, inputs...; kwargs...)

Arguments

f: Function for which the frule should be tested.

inputs either the primal inputs x, or primals and their tangents: x ⊢ ẋ

x: input at which to evaluate f (should generally be set to an arbitary point in the domain).

ẋ: differential w.r.t. x, will be generated automatically if not provided

Non-differentiable arguments, such as indices, should have ẋ set as NoTangent().

Keyword Arguments

output_tangent tangent to test accumulation of derivatives against should be a differential for the output of f. Is set automatically if not provided.

fdm::FiniteDifferenceMethod: the finite differencing method to use.

If check_inferred=true, then the inferrability of the frule is checked, as long as f is itself inferrable.

fkwargs are passed to f as keyword arguments.

All remaining keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_rrule — Method

test_rrule(f, inputs...; kwargs...)

Arguments

f: Function to which rule should be applied.

inputs either the primal inputs x, or primals and their tangents: x ⊢ ẋ

x: input at which to evaluate f (should generally be set to an arbitary point in the domain).

x̄: currently accumulated cotangent, will be generated automatically if not provided

Non-differentiable arguments, such as indices, should have x̄ set as NoTangent().

Keyword Arguments

output_tangent the seed to propagate backward for testing (techncally a cotangent). should be a differential for the output of f. Is set automatically if not provided.

fdm::FiniteDifferenceMethod: the finite differencing method to use.

If check_inferred=true, then the inferrability of the rrule is checked — if f is itself inferrable — along with the inferrability of the pullback it returns.

fkwargs are passed to f as keyword arguments.

All remaining keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_scalar — Method

test_scalar(f, z; rtol=1e-9, atol=1e-9, fdm=central_fdm(5, 1), fkwargs=NamedTuple(), check_inferred=true, kwargs...)

Given a function f with scalar input and scalar output, perform finite differencing checks, at input point z to confirm that there are correct frule and rrules provided.

Arguments

f: Function for which the frule and rrule should be tested.

z: input at which to evaluate f (should generally be set to an arbitary point in the domain).

fkwargs are passed to f as keyword arguments. If check_inferred=true, then the type-stability of the frule and rrule are checked. All remaining keyword arguments are passed to isapprox.