API Documentation

ChainRulesTestUtils.:⊢ — Type

primal ⊢ tangent

Infix shorthand method to construct a PrimalAndTangent. Enter via \vdash + tab on supporting editors.

ChainRulesTestUtils.TestIterator — Type

TestIterator{T,IS<:Base.IteratorSize,IE<:Base.IteratorEltype}

A configurable iterator for testing purposes.

TestIterator(data, itersize, itereltype)
TestIterator(data)

The iterator wraps another iterator data, such as an array, that must have at least as many features implemented as the test iterator and have a FiniteDifferences.to_vec overload. By default, the iterator it has the same features as data.

The optional methods eltype, length, and size are automatically defined and forwarded to data if the type arguments indicate that they should be defined.

ChainRulesTestUtils.test_approx — Function

test_approx(actual, expected, [msg]; kwargs...)

@test's that actual ≈ expected, but breaks up data such that human readable results are shown on failures. Understands things like unthunking ChainRuleCore.Thunks, etc.

If provided msg is printed on a failure. Often additional items are appended to msg to give bread-crumbs into nested structures.

All keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_frule — Method

test_frule(f, args..; kwargs...)

Arguments

f: Function for which the frule should be tested. Can also provide f ⊢ ḟ.

args either the primal args x, or primals and their tangents: x ⊢ ẋ

x: input at which to evaluate f (should generally be set to an arbitary point in the domain).

ẋ: differential w.r.t. x, will be generated automatically if not provided

Non-differentiable arguments, such as indices, should have ẋ set as NoTangent().

Keyword Arguments

output_tangent tangent to test accumulation of derivatives against should be a differential for the output of f. Is set automatically if not provided.

fdm::FiniteDifferenceMethod: the finite differencing method to use.

frule_f=frule: Function with an frule-like API that is tested (defaults to frule). Used for testing gradients from AD systems.

If check_inferred=true, then the inferrability of the frule is checked, as long as f is itself inferrable.

fkwargs are passed to f as keyword arguments.

All remaining keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_rrule — Method

test_rrule(f, args...; kwargs...)

Arguments

f: Function to which rule should be applied. Can also provide f ⊢ f̄.

args either the primal args x, or primals and their tangents: x ⊢ x̄

x: input at which to evaluate f (should generally be set to an arbitary point in the domain).

x̄: currently accumulated cotangent, will be generated automatically if not provided

Non-differentiable arguments, such as indices, should have x̄ set as NoTangent().

Keyword Arguments

output_tangent the seed to propagate backward for testing (technically a cotangent). should be a differential for the output of f. Is set automatically if not provided.

fdm::FiniteDifferenceMethod: the finite differencing method to use.

rrule_f=rrule: Function with an rrule-like API that is tested (defaults to rrule). Used for testing gradients from AD systems.

If check_inferred=true, then the inferrability of the rrule is checked — if f is itself inferrable — along with the inferrability of the pullback it returns.

fkwargs are passed to f as keyword arguments.

All remaining keyword arguments are passed to isapprox.

ChainRulesTestUtils.test_scalar — Method

test_scalar(f, z; rtol=1e-9, atol=1e-9, fdm=central_fdm(5, 1), fkwargs=NamedTuple(), check_inferred=true, kwargs...)

Given a function f with scalar input and scalar output, perform finite differencing checks, at input point z to confirm that there are correct frule and rrules provided.

Arguments

f: Function for which the frule and rrule should be tested.

z: input at which to evaluate f (should generally be set to an arbitary point in the domain).

fkwargs are passed to f as keyword arguments. If check_inferred=true, then the type-stability of the frule and rrule are checked. All remaining keyword arguments are passed to isapprox.

ChainRulesTestUtils.@maybe_inferred — Macro

@maybe_inferred [Type] f(...)

Like @inferred, but does not check the return type if tests are run as part of PkgEval or if the environment variable CHAINRULES_TEST_INFERRED is set to false.